requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-genai==1.11.1
google-api-python-client==2.149.0
numpy==1.26.4

